Resolve flakyness in linearity test
In rare numerical corner cases, the linearity tests report a false
positive as it gets a wrong assumption on which part of the saturation
range has constant saturation step length. This patch relaxes the
linearity test by always skipping the last two rows instead of only the
last row.

In the false positives that spurred this change, the last saturation
step in `right_lin_seg` is double the size of the other saturation
steps. This is most likely a consequence of how the saturation points
are chosen based on the floating point input, and constitutes the corner
case.